requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
dj-database-url==0.5.0
Django==3.0.3
django-admin-shortcuts==2.0.0
django-colorfield==0.3.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-simple-history==2.10.0
django-username-email==2.3.0
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
wincertstore==0.2
